requirements.txt
PyQt5==5.15.9
torch==2.0.1
torchvision==0.15.2
numpy==1.24.4
matplotlib==3.7.2
pydicom==2.4.3
Pillow==10.0.0